remove DynamicClassLoader
